code refactoring Relates #35 close #35
