+++ src/components/Dropzone/index.jsx
@@ -1,0 +1,67 @@
+import React, { useMemo } from "react";
+import { useDropzone } from "react-dropzone";
+
+const baseStyle = {
+  flex: 1,
+  display: "flex",
+  flexDirection: "column",
+  alignItems: "center",
+  padding: "20px",
+  borderWidth: 2,
+  borderRadius: 2,
+  borderColor: "#eeeeee",
+  borderStyle: "dashed",
+  backgroundColor: "#fafafa",
+  color: "#bdbdbd",
+  outline: "none",
+  height: "100px",
+  transition: "border .24s ease-in-out"
+};
+
+const activeStyle = {
+  borderColor: "yellow"
+};
+
+const acceptStyle = {
+  borderColor: "#00e676"
+};
+
+const rejectStyle = {
+  borderColor: "#ff1744"
+};
+
+export default function StyledDropzone(props) {
+  const {
+    getRootProps,
+    getInputProps,
+    isDragActive,
+    isDragAccept,
+    isDragReject
+  } = useDropzone({
+    accept: "image/*",
+    onDrop: files => {
+      if (Array.isArray(files) && files.length)
+        props.getSignedRequest(files[0]);
+    }
+  });
+
+  const style = useMemo(
+    () => ({
+      ...baseStyle,
+      ...(isDragActive ? activeStyle : {}),
+      ...(isDragAccept ? acceptStyle : {}),
+      ...(isDragReject ? rejectStyle : {})
+    }),
+    [isDragAccept, isDragActive, isDragReject]
+  );
+
+  return (
+    <div className="dropzone-container">
+      <div {...getRootProps({ style })}>
+        <input {...getInputProps()} />
+        <div>{getInputProps}</div>
+        <p>Drag 'n' drop some files here, or click to select files</p>
+      </div>
+    </div>
+  );
+}
