--- src/components/UploadForm/stylesVarsHooks.jsx
@@ -1,70 +1,0 @@
-const baseStyle = {
-  flex: 1,
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "center",
-  padding: "20px",
-  borderWidth: 2,
-  borderRadius: 2,
-  borderColor: "#eeeeee",
-  borderStyle: "dashed",
-  backgroundColor: "#fafafa",
-  color: "#bdbdbd",
-  outline: "none",
-  height: "100px",
-  transition: "border .24s ease-in-out"
-};
-
-const activeStyle = {
-  borderColor: "yellow"
-};
-
-const acceptStyle = {
-  borderColor: "#00e676"
-};
-
-const rejectStyle = {
-  borderColor: "#ff1744"
-};
-
-const thumbsContainer = {
-  display: "flex",
-  flexDirection: "row",
-  flexWrap: "wrap",
-  marginTop: 16,
-  justifyContent: "center"
-};
-
-const thumb = {
-  display: "inline-flex",
-  borderRadius: 2,
-  border: "1px solid #eaeaea",
-  marginBottom: 8,
-  marginRight: 8,
-  width: 100,
-  height: 100,
-  padding: 4,
-  boxSizing: "border-box"
-};
-
-const thumbInner = {
-  display: "flex",
-  minWidth: 0,
-  overflow: "hidden"
-};
-
-const img = {
-  display: "block",
-  width: "auto",
-  height: "100%"
-};
-export {
-  baseStyle,
-  activeStyle,
-  acceptStyle,
-  rejectStyle,
-  thumbsContainer,
-  thumb,
-  thumbInner,
-  img
-};
